requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
cryptography==46.0.3
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
google-auth==2.41.1
graphviz==0.21
identify==2.6.15
idna==3.11
nodeenv==1.9.1
platformdirs==4.5.0
pre_commit==4.4.0
psycopg==3.2.11
psycopg-binary==3.2.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.5
python-decouple==3.8
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
virtualenv==20.35.4
